Move and clean up TileStates
