+++ triforce/zelda_game.py
@@ -3,8 +3,8 @@ from enum import Enum
 import numpy as np
 
 from .zelda_enums import AnimationState
-from .zelda_game_state import ID_MAP, ITEM_MAP, ZeldaEnemyId, Direction, ZeldaItemId, tile_index_to_position, \
-        position_to_tile_index
+from .zelda_enums import ID_MAP, ITEM_MAP, ZeldaEnemyId, ZeldaItemId, Direction
+from .tile_states import tile_index_to_position, position_to_tile_index
 from .zelda_game_data import zelda_game_data
 
 MODE_REVEAL = 3
@@ -144,52 +144,6 @@ def is_sword_frozen(state):
 
     return x <= 0x10 or x >= 0xd9 or y <= 0x53 or y >= 0xc5
 
-def init_walkable_tiles():
-    """Returns a lookup table of whether particular tile codes are walkable."""
-    tiles = [0x26, 0x24, 0xf3, 0x8d, 0x91, 0xac, 0xad, 0xcc, 0xd2, 0xd5, 0x68, 0x6f, 0x82, 0x78, 0x7d]
-    tiles += [0x84, 0x85, 0x86, 0x87]
-    tiles += list(range(0x74, 0x77+1))  # dungeon floor tiles
-    tiles += list(range(0x98, 0x9b+1))  # dungeon locked door north
-    tiles += list(range(0xa4, 0xa7+1))  # dungeon locked door east
-
-    return tiles
-
-def init_half_walkable_tiles():
-    """Returns tiles that the top half of link can pass through."""
-    return [0x95, 0x97, 0xb1, 0xb3, 0xd5, 0xd7, 0xc5, 0xc7, 0xc9, 0xcb, 0xd4, 0xb5, 0xb7,
-            0xaf, 0xb9, 0xbb, 0xad, 0xb1, 0xdd, 0xde, 0xd9, 0xdb, 0xdf, 0xd1, 0xdc, 0xd0, 0xda
-            ]
-
-WALKABLE_TILES = init_walkable_tiles()
-HALF_WALKABLE_TILES = init_half_walkable_tiles()
-BRICK_TILE = 0xf6
-
-def tiles_to_weights(tiles) -> None:
-    """Converts the tiles from RAM to a set of weights for the A* algorithm."""
-    brick_mask = tiles == BRICK_TILE
-    tiles[brick_mask] = TileState.BRICK.value
-
-    walkable_mask = np.isin(tiles, WALKABLE_TILES)
-    tiles[walkable_mask] = TileState.WALKABLE.value
-
-    half_walkable_mask = np.isin(tiles, HALF_WALKABLE_TILES)
-    tiles[half_walkable_mask] = TileState.HALF_WALKABLE.value
-
-    tiles[~brick_mask & ~walkable_mask & ~half_walkable_mask] = TileState.IMPASSABLE.value
-
-def is_room_loaded(tiles):
-    """Returns True if the room is loaded."""
-    any_walkable = np.isin(tiles, WALKABLE_TILES).any()
-    return any_walkable
-
-class TileState(Enum):
-    """The state of a tile."""
-    HALF_WALKABLE = 99
-    IMPASSABLE = 100
-    WALKABLE = 1
-    WARNING = 2    # tiles next to enemy, or the walls in a wallmaster room
-    DANGER = 3     # enemy or projectile
-    BRICK = 4      # dungeon bricks
 
 class ZeldaSoundsPulse1(Enum):
     """Sound codes for the game."""
@@ -423,7 +377,6 @@ __all__ = [
     'get_heart_halves',
     'get_heart_containers',
     'has_beams',
-    'TileState',
     'position_to_tile_index',
     'tile_index_to_position',
     'is_sword_frozen',
@@ -433,7 +386,5 @@ __all__ = [
     ZeldaEnemyId.__name__,
     ZeldaItemId.__name__,
     AnimationState.__name__,
-    tiles_to_weights.__name__,
-    is_room_loaded.__name__,
     Direction.__name__,
     ]
